refactor: move `get_jmap_client` to `jmap.py`
